Fix configuration_loader
Somehow while rebasing the branch,
the implementation of the allowlist-related
configuration fields where emitted.
